+++ Toggle.cs
@@ -1,0 +1,7 @@
+namespace BreedingUpgrades;
+
+public enum Toggle
+{
+	On = 1,
+	Off = 0
+}